chore: add condition to not break knex migration rollback on dev environment

+++ migrations/20231204201328_create_schema.ts
@@ -12,5 +12,7 @@ export async function up(knex: Knex): Promise<void> {
 }
 
 export async function down(knex: Knex): Promise<void> {
-  await knex.raw(`DROP SCHEMA IF EXISTS "${schema}" CASCADE`)
+  if (schema !== 'public') {
+    await knex.raw(`DROP SCHEMA IF EXISTS "${schema}" CASCADE`)
+  }
 }
